+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- `service` field support, will use it if exists, or will use ENV service name otherwise.
+
+### Changed
+- log S3 key/path now includes `service` and `entity` prefixes
+
 ## [1.2.0] - 2019-10-11
 ### Added
 - target bucket stage is obtained from `JANIS_ENV` environment variable